feat(sass): use preprocessor for better compatibility
I'm using CSS nesting which is a newish feature not support on all
browsers yet. With a preprocessor like SASS it gets compiled to
un-nested CSS, which doesn't bring any compatibility issues.

In the future I might use other SASS features but for the moment that's it.

Ref: 901d0cc
CSS Nesting compatibility: https://caniuse.com/css-nesting