feat: add Sass support by compiling all scss files into css
- feat: add function to inline css compiled files from scss into a template

For better experience, it's better to put all essential css styles
directly in a <style> tag. This way, as soon as the user receives the
html, the essential styles are already included, and the browser doesn't
have to do another request after it has receive the main content. On slow
connection it's more noticeable, for a brief moment the content might not
be styled while the browser is waiting for stylesheets.