feat: post URL based on shortTitle if it exist in frontmatter
When post title is too long, I can add `shortTitle` to the frontmatter
as an alternate title which is used to generate the URL. If `shortTitle`
doesn't exist (e.g post title is already short) then it fallbacks to
original title.

Example:
- Before: /blog/changing-hosting-provider-and-css-improvements
- After: /blog/hosting-and-css-improvements

For this we moved from a json data file (blog.json) to a javascript
data file (blog.11tydata.js) to be able to compute the values for
the URL.

This change removes the trailing slash from from URL. By returning
`.html` at the end. To return a trailing slash replace `.html` for `/`

```
    return `/blog/${slugifyString(data.shortTitle)}.html`
```

Since I want to simplify URLs as much as possible[^1], the trailing .html
is also removed in the eleventy config.

Ref:
- [^1]: https://rafaelmadriz.com/blog/from-zola-to-eleventy#url-and-titles