feat: use grid layout
- Allow elements to overflow/bleed wider than text content with
  grid-bleed. And allow elements to overflow full page width
  grid-full-bleed. This useful for make images wider than text content.

  Ref:
  - I first saw this in here: https://www.jonashietala.se/blog/2023/10/04/giving_the_blog_a_facelift/#post-overflow-images-and-code-blocks
  - This was also helpful: https://www.joshwcomeau.com/css/full-bleed/

This is also some groundwork for adding sidenotes and table of contents
in the future. Now the grid layout is ready for each sidebar. It should
(in theory) be easier when I add them.

- Use HTML for main homepage instead of markdown. Homepage is not going
to have of written content and since I needed to add some wrappers around
it for new layout, HTML it's easier.